disable errors.Is and errors.As in Go 1.12 and before
