+++ aws/awserr/types.go
@@ -2,7 +2,6 @@ package awserr
 
 import (
 	"encoding/hex"
-	"errors"
 	"fmt"
 )
 
@@ -104,27 +103,6 @@ func (b baseError) OrigErr() error {
 		return NewBatchError("BatchedErrors",
 			"multiple errors occurred", b.errs)
 	}
-}
-
-// Is reports whether this error matches target.
-func (b baseError) Is(target error) bool {
-	for _, e := range b.errs {
-		if errors.Is(e, target) {
-			return true
-		}
-	}
-	return false
-}
-
-// As finds the first error in errs that matches target, and if so,
-// sets target to that error value and returns true. Otherwise, it returns false.
-func (b baseError) As(target interface{}) bool {
-	for _, e := range b.errs {
-		if errors.As(e, &target) {
-			return true
-		}
-	}
-	return false
 }
 
 // OrigErrs returns the original errors if one was set. An empty slice is
